Health program
Get the number of students in each class, plus their age, height, and weight and store each in a list. Print the average age, height and weight of each class on a separate line in the output and then print the class with the highest average height in the output.