Object-oriented project
There is a class named Footballer that inherits properties from the class named Human.
Initially, 22 Footballer objects are created and the names of the players are assigned to each object. Using the random and inheritance methods, these 22 names are divided between two teams A and B and finally the name of each player is printed in the output along with the name of the team he is in.